refactor: split AdvancedTab.tsx into 4 focused sub-tab components
- LanguageProfilesTab.tsx: language profile CRUD (~280 LOC)
- LibrarySourcesTab.tsx: standalone mode + watched folders (~280 LOC)
- BackupTab.tsx: backup creation, listing, restore, retention policy (~200 LOC)
- SubtitleToolsTab.tsx: HI removal, timing, fixes, preview tools (~165 LOC)
- AdvancedTab.tsx becomes 4-line barrel re-exporting all components
- All lazy imports in LegacySettings.tsx and LanguageProfiles.tsx continue to work
